refactor(wamuseum-client): :sparkles: add button convention
