+++ apps/wamuseum-client/ui/index.ts
@@ -1,3 +1,4 @@
+export * from './add-button'
 export * from './create-user-request-form'
 export * from './header'
 export * from './login-form'
